+++ src/testing/classtest.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections;
 
 namespace Python.Test
@@ -58,5 +59,16 @@ namespace Python.Test
 
     internal class InternalClass
     {
+    }
+
+    public class SimpleClass
+    {
+        public static void TestObject(object obj)
+        {
+            if ((!(obj is ISayHello1 && obj is SimpleClass)))
+            {
+                throw new Exception("Expected ISayHello and SimpleClass instance");
+            }
+        }
     }
 }