+++ CHANGELOG.md
@@ -16,6 +16,8 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 -   Fixed error occuring when inheriting a class containing a virtual generic method.
 -   Make a second call to `pythonnet.load` a no-op, as it was intended.
 
+- Added support for multiple inheritance when inheriting from a class and/or multiple interfaces.
+
 ## [3.0.1](https://github.com/pythonnet/pythonnet/releases/tag/v3.0.1) - 2022-11-03
 
 ### Added